added API directory and move classes there
